Arrumando as classes

+++ agendador-horarios/src/main/java/com/Luciano/agendador_horarios/service/AgendamentoService.java
@@ -1,7 +1,9 @@
 package com.Luciano.agendador_horarios.service;
 
 import com.Luciano.agendador_horarios.infrastructure.entity.Agendamento;
+import com.Luciano.agendador_horarios.infrastructure.entity.Proprietario;
 import com.Luciano.agendador_horarios.infrastructure.repository.AgendamentoRepository;
+import com.Luciano.agendador_horarios.security.PasswordValidatorService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -16,38 +18,71 @@ import java.util.Objects;
 public class AgendamentoService {
 
     private final AgendamentoRepository agendamentoRepository;
-    private final BCryptPasswordEncoder bCryptPasswordEncoder;
+    private final PasswordValidatorService passwordValidatorService;
+    private final BCryptPasswordEncoder passwordEncoder;
 
     public Agendamento salvarAgendamento(Agendamento agendamento) {
 
         LocalDateTime horaAgendamento = agendamento.getDataHoraAgendamento();
         LocalDateTime horaFim = agendamento.getDataHoraAgendamento().plusMinutes(1);
 
-        Agendamento agendados = agendamentoRepository.findByServicoAndDataHoraAgendamentoBetween(String.valueOf(agendamento.getServico()),
-                horaAgendamento, horaFim);
+        Agendamento agendados =
+                agendamentoRepository.findByServicoAndDataHoraAgendamentoBetween(
+                        String.valueOf(agendamento.getServico())
+                        , horaAgendamento
+                        , horaFim
+                );
 
-
-        if(Objects.nonNull(agendados)){
+        if (Objects.nonNull(agendados)) {
             throw new RuntimeException("Horário já está preenchido");
         }
+
         return agendamentoRepository.save(agendamento);
     }
 
-    public void deletarAgendamento(LocalDateTime dataHoraAgendamento, String cliente){
+    public void deletarAgendamento(
+            LocalDateTime dataHoraAgendamento,
+            String cliente
+    ) {
+
+      Agendamento agendamento = agendamentoRepository.findByCliete(cliente);
+
+       if (Objects.nonNull(agendamento)) {
+           throw new RuntimeException("Agendamento não encontrado!");
+       }
+
         agendamentoRepository.deleteByDataHoraAgendamentoAndCliente(dataHoraAgendamento, cliente);
     }
 
-    public List<Agendamento> buscarAgendamentosDia(LocalDate data){
+    public List<Agendamento> buscarAgendamentosDia(
+            LocalDate data,
+            String cliente
+    ) {
+
+        Agendamento agendamento =
+                agendamentoRepository.findByCliete(cliente);
+
+        if (Objects.nonNull(agendamento)) {
+            throw new RuntimeException("Agendamento não encontrado!");
+        }
+
         LocalDateTime primeiraHoraDia = data.atStartOfDay();
         LocalDateTime horaFinalDia = data.atTime(20, 59, 59);
 
         return agendamentoRepository.findByDataHoraAgendamentoBetween(primeiraHoraDia, horaFinalDia);
     }
 
-    public Agendamento alterarAgendamento(Agendamento agendamento, String cliente, LocalDateTime dataHoraAgendamento){
-        Agendamento agenda = agendamentoRepository.findByDataHoraAgendamentoAndCliente(dataHoraAgendamento, cliente);
+    public Agendamento alterarAgendamento(
+            Agendamento agendamento,
+            String cliente,
+            LocalDateTime dataHoraAgendamento
+    ) {
+        Agendamento agenda = agendamentoRepository.
+                findByDataHoraAgendamentoAndCliente(
+                dataHoraAgendamento, cliente
+        );
 
-        if(Objects.isNull(agenda)){
+        if (Objects.isNull(agenda)) {
             throw new RuntimeException("Horário não está preenchido");
         }
